Moving controller-runtime-main logic into main. Removing experimental controllers functionality
