+++ cmd/controller/main.go
@@ -19,15 +19,21 @@ package main
 import (
 	"context"
 	"flag"
+	"github.com/knative/eventing/pkg/controller/eventing/subscription"
+	"k8s.io/apimachinery/pkg/runtime"
 	"log"
 	"net/http"
+	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"time"
 
 	// Uncomment the following line to load the gcp plugin (only required to authenticate against GKE clusters).
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 
+	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
 	"github.com/knative/eventing/pkg/logconfig"
 	"github.com/knative/eventing/pkg/system"
+	istiov1alpha3 "github.com/knative/pkg/apis/istio/v1alpha3"
 	"github.com/knative/pkg/configmap"
 	"github.com/knative/pkg/logging"
 	"github.com/knative/pkg/logging/logkey"
@@ -36,6 +42,7 @@ import (
 	"go.uber.org/zap"
 	"k8s.io/client-go/kubernetes"
 	controllerruntime "sigs.k8s.io/controller-runtime/pkg/client/config"
+	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
 )
 
 const (
@@ -44,12 +51,18 @@ const (
 )
 
 var (
-	experimentalControllers string
-	hardcodedLoggingConfig  bool
+	hardcodedLoggingConfig bool
 )
+
+// SchemeFunc adds types to a Scheme.
+type SchemeFunc func(*runtime.Scheme) error
+
+// ProvideFunc adds a controller to a Manager.
+type ProvideFunc func(manager.Manager) (controller.Controller, error)
 
 func main() {
 	flag.Parse()
+	logf.SetLogger(logf.ZapLogger(false))
 
 	// Read the logging config and setup a logger.
 	cm := getLoggingConfigOrDie()
@@ -62,7 +75,7 @@ func main() {
 	defer logger.Sync()
 	logger = logger.With(zap.String(logkey.ControllerType, logconfig.Controller))
 
-	logger.Info("Starting the Controller")
+	logger.Info("Starting the controller")
 
 	// set up signals so we handle the first shutdown signal gracefully
 	stopCh := signals.SetupSignalHandler()
@@ -81,13 +94,41 @@ func main() {
 	configMapWatcher := configmap.NewInformedWatcher(kubeClient, system.Namespace)
 	configMapWatcher.Watch(logconfig.ConfigName, logging.UpdateLevelFromConfigMap(logger, atomicLevel, logconfig.Controller, logconfig.Controller))
 	if err = configMapWatcher.Start(stopCh); err != nil {
-		logger.Fatalf("failed to start controller config map watcher: %v", err)
+		logger.Fatalf("Failed to start controller config map watcher: %v", err)
 	}
 
-	// Start the controller-runtime controllers.
+	// Setup a Manager
+	mgr, err := manager.New(cfg, manager.Options{})
+	if err != nil {
+		logger.Fatalf("Failed to create manager: %v", err)
+	}
+
+	// Add custom types to this array to get them into the manager's scheme.
+	schemeFuncs := []SchemeFunc{
+		istiov1alpha3.AddToScheme,
+		eventingv1alpha1.AddToScheme,
+	}
+	for _, schemeFunc := range schemeFuncs {
+		if err := schemeFunc(mgr.GetScheme()); err != nil {
+			logger.Fatalf("Error adding type to manager's scheme: %v", err)
+		}
+	}
+
+	// Add each controller's ProvideController func to this list to have the
+	// manager run it.
+	providers := []ProvideFunc{
+		subscription.ProvideController,
+	}
+	for _, provider := range providers {
+		if _, err := provider(mgr); err != nil {
+			logger.Fatalf("Error adding controller to manager: %v", err)
+		}
+	}
+
+	// Start the Manager
 	go func() {
-		if err := controllerRuntimeStart(logger, experimentalControllers); err != nil {
-			logger.Fatalf("Error running controller-runtime controllers: %v", err)
+		if err := mgr.Start(stopCh); err != nil {
+			logger.Fatalf("Error starting manager: %v", err)
 		}
 	}()
 
@@ -110,7 +151,6 @@ func main() {
 }
 
 func init() {
-	flag.StringVar(&experimentalControllers, "experimentalControllers", "", "List of experimental controllers to include in the Knative Controller.")
 	flag.BoolVar(&hardcodedLoggingConfig, "hardCodedLoggingConfig", false, "If true, use the hard coded logging config. It is intended to be used only when debugging outside a Kubernetes cluster.")
 }
